New Survey Technique
It's not complete yet but the basic code is there. Basically you add each question and when you press add it should add the question name and type to a list, this list will be put into another list which would be stored in the database and called when opening up the survey.